book_translator/requirements.txt
# python 3.11

python-docx==1.1.2
transformers==4.51.3
torch==2.6.0
sentencepiece==0.2.0
sacremoses==0.1.1
tqdm==4.67.1